imageloader and restful components now have disk level cache.

+++ docs/db/bioroid_imageloader.sql
@@ -2,6 +2,6 @@ DROP TABLE IF EXISTS bioroid_imageloader_cache;
 CREATE TABLE bioroid_imageloader_cache ( 
 	_id INTEGER PRIMARY KEY AUTOINCREMENT, 
 	last_accessed_date REAL UNIQUE ON CONFLICT REPLACE,
-	filesize INTEGER,
+	filesize REAL,
 	imagepath TEXT
 );
